refactor: split image galery content handler logic into modules with imperative wrapper functions and pure functions in the modules
